refactor(frontend): rename icons used in conversions
