+++ src/frontend/src/lib/components/icons/IconConvertFrom.svelte
