+++ src/frontend/src/lib/components/icons/IconConvertTo.svelte
